utils.py
import re
import nltk
from nltk.corpus import stopwords
import string

# Ensure stopwords are downloaded
try:
    nltk.data.find('corpora/stopwords')
except LookupError:
    nltk.download('stopwords')

stop_words = set(stopwords.words('english'))

def clean_text(text):
    """
    Cleans the input text by:
    1. Lowercasing
    2. Removing URLs
    3. Removing punctuation
    4. Removing extra whitespace
    5. Removing stopwords
    """
    if not isinstance(text, str):
        return ""
        
    text = text.lower()
    # Remove URLs
    text = re.sub(r'http\S+|www\S+|https\S+', '', text, flags=re.MULTILINE)
    # Remove user mentions or emails (simple regex)
    text = re.sub(r'\S*@\S*\s?', '', text)
    # Remove punctuation
    text = text.translate(str.maketrans('', '', string.punctuation))
    # Remove extra whitespace
    text = re.sub(r'\s+', ' ', text).strip()
    
    # Remove stopwords
    tokens = text.split()
    tokens = [word for word in tokens if word not in stop_words]
    
    return ' '.join(tokens)
