requirements.txt
flask
tensorflow
pandas
numpy
scikit-learn
pdfplumber
python-docx
pytesseract
Pillow
werkzeug
